requirements.txt
alembic==1.0.5
atomicwrites==1.3.0
attrs==19.3.0
awesome-slugify==1.6.5
bcrypt==3.1.5
bleach==3.1.0
certifi==2019.9.11
cffi==1.11.5
Click==7.0
cloudinary==1.18.2
colorama==0.4.1
decorator==4.4.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Markdown==0.3
Flask-Migrate==2.3.1
Flask-MySQL==1.4.0
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
importlib-metadata==1.1.0
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
mock==3.0.5
more-itertools==8.0.0
mysql-connector-python==8.0.14
packaging==19.2
Pillow==5.4.0
pluggy==0.13.1
protobuf==3.6.1
pur==5.3.0
py==1.8.0
pycparser==2.19
PyMySQL==0.9.3
pyparsing==2.4.5
pytest==5.3.1
python-dateutil==2.7.5
python-editor==1.0.3
regex==2019.1.24
six==1.12.0
SQLAlchemy==1.2.15
SQLAlchemy-Utils==0.36.1
Unidecode==0.4.21
urllib3==1.25.7
validators==0.14.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wtf-tinymce==0.1.3
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
zipp==0.6.0
